requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==22.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
deprecation==2.1.0
et-xmlfile==1.1.0
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.16.2
Flask==3.1.0
fonttools==4.61.0
fsspec==2026.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.5.2
jsonschema==4.17.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
nbformat==5.7.0
nest-asyncio==1.6.0
numpy==2.4.3
opencv-python==4.11.0.86
openpyxl==3.1.2
packaging==26.0
paho-mqtt==2.1.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.4
plotly==5.11.0
postgrest==2.28.3
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
pyrsistent==0.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
pyzmq==27.1.0
realtime==2.28.3
requests==2.32.3
rich==14.3.3
Sastrawi==1.0.1
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.0
six==1.17.0
stack-data==0.6.3
starlette==0.52.1
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==8.1.0
threadpoolctl==3.6.0
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.23.0
zstandard==0.25.0
